Move LocalityDetails to separate folder
